requirements.txt
pandas
snowflake-connector-python[pandas]
rapidfuzz
python-dotenv
pandas
numpy
